chore(bazel): add --incompatible_strict_action_env
This is the workaround to solve the issue when `com_google_protobuf`
is recompiling protoc compiler on every bazel build.
As Alex Eagle meantioned this issue could be realated to the order
of `rules_proto`, `com_google_protobuf` and the Gazelle loading
order, because eaach of them could override the other one and
to load it's own internal default proto rules if none is found
at load time.

Resolves: #152
See also: #162