chore(hygen): add hygen
Intention is to generate component boilerplate.
Current issue is that Hygen does not work well in monorepo
setting. At least i was not able the get the flawless workflow
working, because template nesting is not working like i expect it
to work.
Will try to fix it one more time or will look into alternatives.