feat(ansible): move vault init into separate role
Main problem was within "race conditions" and bad control flow.
Because Init was in the same role as provisioning and configuring, when
Terraform was destroying resources it was expecting Vault instance
running.
Now, because init are on separate play, i can easy control the flow.

But i think i will move to baking Vault Gold Image for faster
provisioning. I need to figure out, how to interact with Packer
automatically.