feat(ansible): refactor lua plugins
Main idea was to simplify plugin configuration.
It should be possible just to throw in single file and to enable/disable
and configure the plugin from that file.
Currently i left the "modules" separtation, but i am thinking to make
all those pugins into single flat file structure.
This role now should be broaken because i changed the require()'s.