feat(ansible): create hyprland role
Quite dirty and "just make it to work".
I got it working but it involves of tweaking Qemu Display and Video
settings by enabling OpenGL. It is slow. But for testing it works fine.
Other task is to attach dedicated keyboard to it as i don't know how to
pass other Super key to it.